7/21 home.js/home.css 내용 common에 통합해서 사용 헤더 임시로 css 적용 community.css 일부 수정
