--- src/main/java/com/spring/nong4/indexController.java
@@ -1,12 +1,0 @@
-package com.spring.nong4;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class IndexController {
-
-    @RequestMapping("/index")
-    public void index(){
-    }
-}
