+++ src/main/webapp/WEB-INF/views/temp/tilesHeader.jsp
@@ -3,17 +3,16 @@
 <%@ taglib prefix="sec" uri="http://www.springframework.org/security/tags" %>
 <header>
     <ul>
+        <li><a href="/board/home">메인</a></li>
+        <sec:authorize access="isAuthenticated()">
+            <li><a href="/board/friendBoard">글쓰기</a></li>
+        </sec:authorize>
         <sec:authorize access="isAnonymous()"><!-- -->
         <li><a href="/user/login">로그인</a></li>
         <li><a href="/user/join">회원가입</a></li>
         </sec:authorize>
         <sec:authorize access="isAuthenticated()">
             <li><a href="/user/logout">로그아웃</a></li>
-        </sec:authorize>
-        <li><a href="/board/home">리스트</a></li>
-        <sec:authorize access="isAuthenticated()">
-            <li><a href="/board/friendBoard">글쓰기</a></li>
-            <li><a href="/board/writeMod">글쓰기</a></li>
         </sec:authorize>
     </ul>
 </header>
