+++ src/main/webapp/res/css/board/community.css
@@ -1,7 +1,4 @@
-html, body { margin: 0; padding: 0; }
-.pointer { cursor: pointer; }
-a { text-decoration: none; }
-.background { width: 100%; height: 100%; background-image: url("/res/img/main.png"); display: flex; }
+.background { height: 100vh; background-image: url("/res/img/main.png"); display: flex;  background-repeat : no-repeat; background-size : cover; }
 .background-left { flex-grow: 2; }
 .background-center { background-color: white;  display: flex; flex-grow: 1;}
 .background-right { flex-grow: 2; }
@@ -20,7 +17,8 @@ a { text-decoration: none; }
 .communityboard {}
 .level { height: 33%; }
 .level .swiper-container { height: 100%; width: 900px; }
-.level .swiper-slide { display: flex; justify-content: center; margin-top: 15px; height: 80%; }
+.level .swiper-slide { display: flex; justify-content: center; align-items: center; }
+.level .swiper-slide img { width: 350px; }
 .news { height: 33%; padding: 5px 30px;}
 .news div { padding: 15px 0px; border-bottom: 1px solid #d3d3d3; }
 .todayvideo { margin: 30px; display: flex; }
