+++ src/main/webapp/res/css/board/home.css
@@ -1,4 +1,4 @@
-.junior-class { cursor: pointer;}
-.intermediate-class { cursor: pointer;}
-.advanced-class { cursor: pointer;}
-.master-class { cursor: pointer;}
+/*.junior-class { cursor: pointer;}*/
+/*.intermediate-class { cursor: pointer;}*/
+/*.advanced-class { cursor: pointer;}*/
+/*.master-class { cursor: pointer;}*/
