+++ src/main/webapp/res/css/common.css
@@ -1,0 +1,8 @@
+html, body, section { margin: 0; padding: 0; }
+.pointer { cursor: pointer; }
+a { text-decoration: none; color: black; }
+header ul { list-style-type: none; margin: 0; padding: 0; background-color: #333; }
+header ul:after{ content:''; display: block; clear:both; }
+header ul li { float: left; }
+header ul li a { display: block; color: white; text-align: center; padding: 14px 16px; }
+header ul li a:hover:not(.active) { background-color: #111; }
